Reorganizing SignDy related functions
